chore: moved utils function out
