+++ packages/components/src/select/utils/defaultFilterFunction/defaultFilterFunction.spec.js
@@ -1,0 +1,3 @@
+describe('defaultFilterFunction', () => {
+  it('Add test here', () => {});
+});
