+++ packages/components/src/select/utils/defaultFilterFunction/index.js
@@ -1,0 +1,1 @@
+export { default } from './defaultFilterFunction';
