+++ packages/components/src/select/utils/index.js
@@ -1,0 +1,2 @@
+export { default as stopPropagation } from './stopPropagation';
+export { default as defaultFilterFunction } from './defaultFilterFunction';
