+++ packages/components/src/select/utils/stopPropagation/index.js
@@ -1,0 +1,1 @@
+export { default } from './stopPropagation';
