+++ packages/components/src/select/utils/stopPropagation/stopPropagation.spec.js
@@ -1,0 +1,3 @@
+describe('Stop propagation', () => {
+  it('Add test here', () => {});
+});
